Separate test modules, remove over-importing, create reports domain
