+++ tests/applications/test_applications_router.py
@@ -1,0 +1,128 @@
+from datetime import datetime, timezone
+import bson
+from fastapi.testclient import TestClient
+from mongomock.database import Database as MockMongoDatabase
+from pymongo.database import Database as MongoDatabase
+from pymongo.errors import WriteError
+import pytest
+
+from src.applications.models import ApplicationModel
+from src.config import APPLICATIONS_COLLECTION
+from src.main import app
+
+client = TestClient(app)
+
+class TestCreateApplication:
+    def test_success_min_required_fields(self, mock_mongo_db: MockMongoDatabase):
+        response = client.post("/api/applications", json={
+            "fname": "First",
+            "lname": "Last",
+            "email": "test.user@cti.com"
+        })
+
+        assert response.status_code == 201
+        assert bson.ObjectId.is_valid(response.json()["_id"])
+
+    def test_success_with_extra_fields(self, mock_mongo_db: MockMongoDatabase):
+        response = client.post("/api/applications", json={
+            "fname": "First",
+            "lname": "Last",
+            "email": "test.user@cti.com",
+            "cohort": True,
+            "graduating_year": 2024
+        })
+
+        assert response.status_code == 201
+        assert bson.ObjectId.is_valid(response.json()["_id"])
+        assert response.json()["cohort"]
+        assert response.json()["graduating_year"] == 2024
+
+    def test_failure_missing_required_last_name(self, mock_mongo_db: MockMongoDatabase):
+        response = client.post("/api/applications", json={
+            "fname": "First",
+            # missing lname
+            "email": "test.user@cti.com",
+            "cohort": True,
+            "graduating_year": 2024
+        })
+
+        assert response.status_code == 422
+        detail = response.json()["detail"][0]
+        assert detail["type"] == "missing"
+
+    def test_failure_invalid_email(self, mock_mongo_db: MockMongoDatabase):
+        response = client.post("/api/applications", json={
+            "fname": "First",
+            "lname": "Last",
+            "email": "test.user@cti",
+            "cohort": True,
+            "graduating_year": 2024
+        })
+
+        assert response.status_code == 422
+        detail = response.json()["detail"][0]
+        assert detail["type"] == "value_error"
+        assert "not a valid email address" in detail["msg"]
+
+    def test_failure_duplicate_email_key(self, mock_mongo_db: MockMongoDatabase):
+        # pydantic model used to validate test data before inserting as mongomock does not support json schema validation
+        app = ApplicationModel(
+            fname="First",
+            lname="Last",
+            email="test.user@cti.com", # email is a unique index
+            cohort=True,
+            graduating_year=2024,
+            app_submitted=datetime.now(timezone.utc)
+        )
+        
+        mock_mongo_db.get_collection(APPLICATIONS_COLLECTION).insert_one(app.model_dump())
+        
+        response = client.post("/api/applications", json={
+            "fname": "First",
+            "lname": "Last",
+            "email": "test.user@cti.com",
+            "cohort": True,
+            "graduating_year": 2024
+        })
+
+        assert response.status_code == 409
+        detail = response.json()["detail"]
+        assert "Duplicate index key" in detail
+
+    @pytest.mark.integration
+    def test_application_persistence(self, real_mongo_db: MongoDatabase):
+        """Integration test validating that a document can be inserted and found
+        if the fields follow the collection schema
+        """
+        app_collection = real_mongo_db.get_collection(APPLICATIONS_COLLECTION)
+        prev_count = app_collection.count_documents({})
+
+        response = client.post("/api/applications", json={
+            "fname": "First",
+            "lname": "Last",
+            "email": "test.user@cti.com"
+        })
+
+        inserted = app_collection.find_one({"email": "test.user@cti.com"})
+
+        assert response.status_code == 201
+        assert bson.ObjectId.is_valid(response.json()["_id"])
+        assert inserted is not None and str(inserted["_id"]) == response.json()["_id"]
+        assert prev_count + 1 == app_collection.count_documents({})
+    
+    @pytest.mark.integration
+    def test_application_schema_rejects_invalid_insert(self, real_mongo_db: MongoDatabase):
+        """Integration test validating that a document will not be inserted
+        if it does not follow the collection json validation schema.
+        
+        Relevant for inserts performed outside of API.
+        """
+        app_collection = real_mongo_db.get_collection(APPLICATIONS_COLLECTION)
+
+        with pytest.raises(WriteError, match="Document failed validation"):
+            app_collection.insert_one({
+                "fname": "First",
+                "lname": "Last",
+                "email": "test.user@cti.com"
+                # missing app_submitted
+            })
